Make reading config a plugin
